feat(passkey): add to multisig
