+++ packages/typescript/test/unit/cryptography/passkey.test.ts
@@ -3,160 +3,22 @@
 
 import { secp256r1 } from '@noble/curves/p256';
 import { blake2b } from '@noble/hashes/blake2b';
-import { sha256 } from '@noble/hashes/sha256';
 import { describe, expect, it } from 'vitest';
 
 import { bcs } from '../../../src/bcs';
 import { messageWithIntent } from '../../../src/cryptography';
 import { PasskeyKeypair } from '../../../src/keypairs/passkey';
-import { findCommonPublicKey, PasskeyProvider } from '../../../src/keypairs/passkey/keypair';
+import { findCommonPublicKey } from '../../../src/keypairs/passkey/keypair';
 import {
 	parseSerializedPasskeySignature,
 	PasskeyPublicKey,
-	SECP256R1_SPKI_HEADER,
 } from '../../../src/keypairs/passkey/publickey';
-import {
-	AuthenticationCredential,
-	RegistrationCredential,
-} from '../../../src/keypairs/passkey/types';
 import { fromBase64 } from '../../../src/utils';
-
-function compressedPubKeyToDerSPKI(compressedPubKey: Uint8Array): Uint8Array {
-	// Combine header with the uncompressed public key coordinates.
-	const uncompressedPubKey = secp256r1.ProjectivePoint.fromHex(compressedPubKey).toRawBytes(false);
-	return new Uint8Array([...SECP256R1_SPKI_HEADER, ...uncompressedPubKey]);
-}
-
-class MockPasskeySigner implements PasskeyProvider {
-	private sk: Uint8Array;
-	private authenticatorData: Uint8Array;
-	private pk: Uint8Array | null;
-	private changeDigest: boolean;
-	private changeClientDataJson: boolean;
-	private changeAuthenticatorData: boolean;
-	private changeSignature: boolean;
-
-	constructor(options?: {
-		sk?: Uint8Array;
-		pk?: Uint8Array;
-		authenticatorData?: Uint8Array;
-		changeDigest?: boolean;
-		changeClientDataJson?: boolean;
-		changeAuthenticatorData?: boolean;
-		changeSignature?: boolean;
-	}) {
-		this.sk = options?.sk ?? secp256r1.utils.randomPrivateKey();
-		this.pk = options?.pk ?? null;
-		this.authenticatorData =
-			options?.authenticatorData ??
-			new Uint8Array([
-				88, 14, 103, 167, 58, 122, 146, 250, 216, 102, 207, 153, 185, 74, 182, 103, 89, 162, 151,
-				100, 181, 113, 130, 31, 171, 174, 46, 139, 29, 123, 54, 228, 29, 0, 0, 0, 0,
-			]);
-		this.changeDigest = options?.changeDigest ?? false;
-		this.changeClientDataJson = options?.changeClientDataJson ?? false;
-		this.changeAuthenticatorData = options?.changeAuthenticatorData ?? false;
-		this.changeSignature = options?.changeSignature ?? false;
-	}
-
-	async create(): Promise<RegistrationCredential> {
-		const pk = this.pk;
-		const credentialResponse: AuthenticatorAttestationResponse = {
-			attestationObject: new Uint8Array().slice().buffer,
-			clientDataJSON: new TextEncoder()
-				.encode(
-					JSON.stringify({
-						type: 'webauthn.create',
-						challenge: '',
-						origin: 'https://www.sui.io',
-						crossOrigin: false,
-					}),
-				)
-				.slice().buffer,
-			getPublicKey: () =>
-				pk
-					? compressedPubKeyToDerSPKI(pk).slice().buffer
-					: new Uint8Array([
-							48, 89, 48, 19, 6, 7, 42, 134, 72, 206, 61, 2, 1, 6, 8, 42, 134, 72, 206, 61, 3, 1, 7,
-							3, 66, 0, 4, 232, 238, 71, 180, 129, 19, 164, 11, 106, 184, 25, 185, 136, 226, 178,
-							64, 72, 105, 218, 94, 85, 28, 244, 5, 19, 172, 167, 65, 137, 42, 193, 31, 97, 55, 49,
-							168, 234, 185, 163, 251, 162, 235, 213, 185, 116, 178, 194, 7, 128, 238, 255, 59, 121,
-							255, 175, 188, 137, 89, 147, 168, 103, 128, 97, 52,
-						]).slice().buffer,
-			getPublicKeyAlgorithm: () => -7,
-			getTransports: () => ['usb', 'ble', 'nfc'],
-			getAuthenticatorData: () => this.authenticatorData.slice().buffer,
-		};
-
-		const credential = {
-			id: 'mock-credential-id',
-			rawId: new Uint8Array([1, 2, 3]).buffer,
-			response: credentialResponse,
-			type: 'public-key',
-			authenticatorAttachment: 'cross-platform',
-			getClientExtensionResults: () => ({}),
-		};
-
-		return credential as RegistrationCredential;
-	}
-
-	async get(challenge: Uint8Array): Promise<AuthenticationCredential> {
-		// Manually mangle the digest bytes if changeDigest.
-		if (this.changeDigest) {
-			challenge = sha256(challenge);
-		}
-
-		const clientDataJSON = this.changeClientDataJson
-			? JSON.stringify({
-					type: 'webauthn.create', // Wrong type for clientDataJson.
-					challenge: Buffer.from(challenge).toString('base64'),
-					origin: 'https://www.sui.io',
-					crossOrigin: false,
-				})
-			: JSON.stringify({
-					type: 'webauthn.get',
-					challenge: Buffer.from(challenge).toString('base64'),
-					origin: 'https://www.sui.io',
-					crossOrigin: false,
-				});
-
-		// Sign authenticatorData || sha256(clientDataJSON).
-		const dataToSign = new Uint8Array([
-			...this.authenticatorData,
-			...sha256(new TextEncoder().encode(clientDataJSON)),
-		]);
-
-		// Manually mangle the signature if changeSignature.
-		const signature = this.changeSignature
-			? secp256r1.sign(sha256(dataToSign), secp256r1.utils.randomPrivateKey())
-			: secp256r1.sign(sha256(dataToSign), this.sk);
-
-		const authResponse: AuthenticatorAssertionResponse = {
-			authenticatorData: this.changeAuthenticatorData
-				? new Uint8Array([1]).buffer // Change authenticator data
-				: this.authenticatorData.slice().buffer,
-			clientDataJSON: new TextEncoder().encode(clientDataJSON).slice().buffer,
-			signature: signature.toDERRawBytes().slice().buffer,
-			userHandle: null,
-		};
-
-		const credential = {
-			id: 'mock-credential-id',
-			rawId: new Uint8Array([1, 2, 3]).buffer,
-			type: 'public-key',
-			response: authResponse,
-			authenticatorAttachment: 'cross-platform',
-			getClientExtensionResults: () => ({}),
-		};
-
-		return credential as AuthenticationCredential;
-	}
-}
+import { createMockPasskeyKeypair, MockPasskeySigner } from './test-utils';
 
 describe('passkey signer E2E testing', () => {
 	it('should retrieve the correct sui address', async () => {
-		const mockProvider = new MockPasskeySigner();
-		const signer = await PasskeyKeypair.getPasskeyInstance(mockProvider);
+		const signer = await createMockPasskeyKeypair();
 		const publicKey = signer.getPublicKey();
 		expect(publicKey.toSuiAddress()).toEqual(
 			'0x05d52348e3e3a785e1e458ebe74d71e21dd4db2ba3088484cab22eca5a07da02',
@@ -170,12 +32,11 @@ describe('passkey signer E2E testing', () => {
 			88, 14, 103, 167, 58, 122, 146, 250, 216, 102, 207, 153, 185, 74, 182, 103, 89, 162, 151, 100,
 			181, 113, 130, 31, 171, 174, 46, 139, 29, 123, 54, 228, 29, 0, 0, 0, 0,
 		]);
-		const mockProvider = new MockPasskeySigner({
+		const signer = await createMockPasskeyKeypair({
 			sk: sk,
 			pk: pk,
 			authenticatorData: authenticatorData,
 		});
-		const signer = await PasskeyKeypair.getPasskeyInstance(mockProvider);
 		const testMessage = new TextEncoder().encode('Hello world!');
 		const { signature } = await signer.signPersonalMessage(testMessage);
 
@@ -209,12 +70,11 @@ describe('passkey signer E2E testing', () => {
 			88, 14, 103, 167, 58, 122, 146, 250, 216, 102, 207, 153, 185, 74, 182, 103, 89, 162, 151, 100,
 			181, 113, 130, 31, 171, 174, 46, 139, 29, 123, 54, 228, 29, 0, 0, 0, 0,
 		]);
-		const mockProvider = new MockPasskeySigner({
+		const signer = await createMockPasskeyKeypair({
 			sk: sk,
 			pk: pk,
 			authenticatorData: authenticatorData,
 		});
-		const signer = await PasskeyKeypair.getPasskeyInstance(mockProvider);
 
 		const messageBytes = fromBase64(
 			'AAABACACAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgEBAQABAABnEUWt6SNz7OPa4hXLyCw9tI5Y7rNxhh5DFljH1jLT6QEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAIMqiyOLCIblSqii0TkS8PjMoj3tmA7S24hBMyonz2Op/ZxFFrekjc+zj2uIVy8gsPbSOWO6zcYYeQxZYx9Yy0+noAwAAAAAAAICWmAAAAAAAAA==',
@@ -245,58 +105,48 @@ describe('passkey signer E2E testing', () => {
 		expect(parsed.clientDataJson).toEqual(clientDataJSONString);
 
 		// Case 1: passkey returns a signature on wrong digest, fails to verify.
-		const mockProviderWrongDigest = new MockPasskeySigner({
+		const signerWrongDigest = await createMockPasskeyKeypair({
 			sk: sk,
 			pk: pk,
 			authenticatorData: authenticatorData,
 			changeDigest: true,
 		});
-		const signerWrongDigest = await PasskeyKeypair.getPasskeyInstance(mockProviderWrongDigest);
 
 		const { signature: wrongSignature } = await signerWrongDigest.signTransaction(messageBytes);
 		isValid = await publicKey.verifyTransaction(messageBytes, wrongSignature);
 		expect(isValid).toBe(false);
 
 		// Case 2: passkey returns wrong type on client data json, fails to verify.
-		const mockProviderWrongClientDataJson = new MockPasskeySigner({
+		const signerWrongClientDataJson = await createMockPasskeyKeypair({
 			sk: sk,
 			pk: pk,
 			authenticatorData: authenticatorData,
 			changeClientDataJson: true,
 		});
-		const signerWrongClientDataJson = await PasskeyKeypair.getPasskeyInstance(
-			mockProviderWrongClientDataJson,
-		);
 		const { signature: wrongSignature2 } =
 			await signerWrongClientDataJson.signTransaction(intentMessage);
 		isValid = await publicKey.verifyTransaction(messageBytes, wrongSignature2);
 		expect(isValid).toBe(false);
 
 		// Case 3: passkey returns mismatched authenticator data, fails to verify.
-		const mockProviderWrongAuthenticatorData = new MockPasskeySigner({
+		const signerWrongAuthenticatorData = await createMockPasskeyKeypair({
 			sk: sk,
 			pk: pk,
 			authenticatorData: authenticatorData,
 			changeAuthenticatorData: true,
 		});
-		const signerWrongAuthenticatorData = await PasskeyKeypair.getPasskeyInstance(
-			mockProviderWrongAuthenticatorData,
-		);
 		const { signature: wrongSignature3 } =
 			await signerWrongAuthenticatorData.signTransaction(intentMessage);
 		isValid = await publicKey.verifyTransaction(messageBytes, wrongSignature3);
 		expect(isValid).toBe(false);
 
 		// Case 4: passkey returns a signature from a mismatch secret key, fails to verify.
-		const mockProviderWrongSignature = new MockPasskeySigner({
+		const signerWrongSignature = await createMockPasskeyKeypair({
 			sk: sk,
 			pk: pk,
 			authenticatorData: authenticatorData,
 			changeSignature: true,
 		});
-		const signerWrongSignature = await PasskeyKeypair.getPasskeyInstance(
-			mockProviderWrongSignature,
-		);
 		const { signature: wrongSignature4 } =
 			await signerWrongSignature.signTransaction(intentMessage);
 		isValid = await publicKey.verifyTransaction(messageBytes, wrongSignature4);
@@ -337,16 +187,19 @@ describe('passkey signer E2E testing', () => {
 		const sk = secp256r1.utils.randomPrivateKey();
 		const pk = secp256r1.getPublicKey(sk);
 		const authenticatorData = new Uint8Array([]);
+		const signer = await createMockPasskeyKeypair({
+			sk: sk,
+			pk: pk,
+			authenticatorData: authenticatorData,
+		});
+		const address = signer.getPublicKey().toSuiAddress();
+
+		const testMessage = new TextEncoder().encode('Hello world!');
 		const mockProvider = new MockPasskeySigner({
 			sk: sk,
 			pk: pk,
 			authenticatorData: authenticatorData,
 		});
-
-		const signer = await PasskeyKeypair.getPasskeyInstance(mockProvider);
-		const address = signer.getPublicKey().toSuiAddress();
-
-		const testMessage = new TextEncoder().encode('Hello world!');
 		const possiblePks = await PasskeyKeypair.signAndRecover(mockProvider, testMessage);
 
 		const testMessage2 = new TextEncoder().encode('Hello world 2!');
